At this point parsing data from uniflow should be possible
