+++ src/bill2myprint/forms.py
@@ -1,1 +1,3 @@
+# -*- coding:utf-8 -*-
+
 from django.forms import ModelForm
