+++ src/bill2myprint/templates/hello_world/login_out.html
@@ -3,7 +3,6 @@
 {% block body %}
     <h1>{{ text }}</h1>
     {% if request.user.is_authenticated %}
-        <a href="{% url 'build2myprint:list_shoes' %}">Liste des chaussures</a>
         <a href="{% url 'logout' %}">Logout</a>
     {% else %}
         <a href="{% url 'login_view' %}">Login</a>
