+++ src/config/dbrouter.py
@@ -7,8 +7,10 @@ class MyPrintRouter(object):
         """
         Attempts to read build2myprint models go to MyPrint MSSQL DB.
         """
-        if model._meta.app_label == 'build2myprint':
+        if model._meta.app_label == 'uniflow':
             return 'myprint'
+        elif model._meta.app_label == 'equitrac':
+            return 'equitrac_transactions'
         return 'default'
 
     def db_for_write(self, model, **hints):
@@ -21,9 +23,12 @@ class MyPrintRouter(object):
         """
         Allow relations if both models are in the build2myprint app.
         """
-        if obj1._meta.app_label == 'build2myprint' and \
-           obj2._meta.app_label == 'build2myprint':
-           return True
+        if obj1._meta.app_label == 'uniflow' and \
+                obj2._meta.app_label == 'uniflow':
+            return True
+        elif obj1._meta.app_label == 'equitrac' and \
+                obj2._meta.app_label == 'equitrac':
+            return 'equitrac_transactions'
         return None
 
     def allow_migrate(self, db, app_label, model_name=None, **hints):
@@ -31,6 +36,7 @@ class MyPrintRouter(object):
         Make sure the build2myprint models are not migrated nor managed by django in DB,
         as th DB already exists.
         """
-        if app_label == 'build2myprint' or db == 'myprint':
+        if app_label == 'uniflow' or db == 'myprint' or \
+                app_label == 'equitrac' or db == 'equitrac_transactions':
             return False
         return None
