+++ src/equitrac/management/commands/get_all_students_from_equitrac.py
@@ -1,0 +1,20 @@
+import re
+
+from django.db.models import Q
+from django.core.management.base import BaseCommand
+
+from equitrac.models import TAllTransactions
+from bill2myprint.models import Section, Student
+
+
+class Command(BaseCommand):
+    help = 'Populates the database with sections and the faculties'
+
+    def handle(self, *args, **options):
+        students = TAllTransactions.objects.filter(hierarchie2='EPFL ETU').exclude(
+            Q(hierarchie3__icontains='audit') | Q(hierarchie3__icontains='EME') | Q(hierarchie3__icontains='EDOC'))
+        students = students.values_list('person_sciper', 'hierarchie3').distinct()
+        for student in students:
+            section_acronym = re.search('EPFL ETU (.*)-.*', student[1]).group(1)
+            section = Section.objects.get(acronym=section_acronym)
+            Student.objects.create(sciper=student[0], section=section)
